Food Data

This dataset contains information regarding food from students of Mercyhurst University

Usage

food_data

Format

A data frame with 125 rows and 3 variables:

comfort_food

Character, list of comfort foods

food_childhood

Character, list of childhood foods

meals_dinner_friend

Character, list of foods eaten when friends came over

Source

https://www.kaggle.com/datasets/borapajo/food-choices/versions/1/data