Mouse Body Weight Data (Clean)

This dataset contains cleaned information for the weight of each mouse taken over three dates.

Usage

mouse_data_bw_clean

Format

A data frame with 32 rows and 7 variables:

ID

Chatracter, description of the mouse sex and number

Body Weight 1

Numeric, the Body Weight of the mouse at Date Body Weight 1

Date Body Weight 1

Numeric, the date Body Weight 1 was taken in POSIXlt format

Body Weight 2

Numeric, the Body Weight of the mouse at Date Body Weight 2

Date Body Weight 2

Numeric, the date Body Weight 2 was taken in POSIXlt format

Body Weight 3

Numeric, the Body Weight of the mouse at Date Body Weight 3

Date Body Weight 3

Numeric, the date Body Weight 3 was taken in POSIXlt format

Source

James Xenakis's Mouse Data