Plot Outliers

This function takes a longitunal dataset and finds any unusually large percent changes between datapoints in each row.

Usage

plot_outliers(data, col_names, threshold = 0.2, ignore_NA = FALSE)

Arguments

data

The name of the dataset

col_names

A vector of column names that

threshold

A number representing the percent change to classify as an "outlier" (default .2, or 20% change)

ignore_NA

A boolean determining whether to ignore NA values

Value

A plot displaying each outlier plotted on a graph

Examples

plot_outliers(mouse_data_bw_clean, (c("Body Weight 1", "Body Weight 2", "Body Weight 3")), .2)