Analyze String Typos

Creates a table with rows that have different values than the rest of the other rows.

Usage

table_different(data, col_name, frequency = 5)

Arguments

data

dataframe The dataframe of information

col_name

character The name of column

frequency

numeric The maximum number of times a string can appear before it no longer appears on the table

Value

Returns a table with a list of row numbers and strings that showed up a fewer number of times than frequency

Examples

table_different(mouse_data_birth, "Treatment", 5)

  

Inconsistent Strings
    

Row Number
      	String
    

11
	Placobo

20
	Trmt